Refactor de las ubicaciones de los tests. Ahora su path es el mismo que la clase que testean.
