+++ composeApp/src/desktopTest/kotlin/data/local/LocalCacheTest.kt
@@ -1,4 +1,4 @@
-package repositoryTest
+package data.local
 
 import edu.dyds.movies.data.local.LocalCache
 import edu.dyds.movies.domain.entity.Movie
